Made ExternalPrinter as an optional feature. Clippy is happy, test pass, docs added.
