+++ Cargo.toml
@@ -32,7 +32,7 @@ rusqlite = { version = "0.28.0", optional = true, features = ["bundled"] }
 serde_json = { version = "1.0.79", optional = true }
 gethostname = { version = "0.2.3", optional = true }
 thiserror = "1.0.31"
-crossbeam = "0.8.2"
+crossbeam = { version = "0.8.2", optional = true }
 
 [dev-dependencies]
 tempfile = "3.3.0"
@@ -43,4 +43,4 @@ rstest = { version = "0.15.0", default-features = false }
 system_clipboard = ["clipboard"]
 bashisms = []
 sqlite = ["rusqlite", "serde_json", "gethostname"]
-
+external_printer = ["crossbeam"]
